refactor(bench): split native-competitors main.rs under file line limit
The file line limit CI check caps new source files at 350 lines, and benchmarks/native-competitors/src/main.rs landed at 426. Extract the timing engine into bench.rs and the JSON serialization into json.rs as sibling modules (mod.rs is disallowed), leaving main.rs at 287 lines. Pure move: no behavior change; cargo test/clippy/fmt all pass.
